Fix float-graded submissions

+++ openedxscorm/scormxblock.py
@@ -257,7 +257,7 @@ class ScormXBlock(XBlock):
                 self.publish_grade()
                 context.update({"lesson_score": self.lesson_score})
         elif name in ["cmi.core.score.raw", "cmi.score.raw"] and self.has_score:
-            self.lesson_score = int(data.get("value", 0)) / 100.0
+            self.lesson_score = float(data.get("value", 0)) / 100.0
             self.publish_grade()
             context.update({"lesson_score": self.lesson_score})
         else:
